refactor: 면접 생성 팩토리 메소드의 접근 제어자를 protected로 변경
