+++ src/main/java/com/mock/interview/interview/application/InterviewService.java
@@ -7,14 +7,13 @@ import com.mock.interview.category.infra.JobPositionRepository;
 import com.mock.interview.experience.application.helper.ExperienceFinder;
 import com.mock.interview.experience.infra.ExperienceRepository;
 import com.mock.interview.interview.application.dto.InterviewTopicDto;
-import com.mock.interview.interview.domain.InterviewStartService;
+import com.mock.interview.interview.domain.model.InterviewStartService;
 import com.mock.interview.interview.domain.model.*;
 import com.mock.interview.interview.presentation.dto.InterviewAccountForm;
 import com.mock.interview.interview.presentation.dto.InterviewConfigForm;
 import com.mock.interview.category.domain.model.JobCategory;
 import com.mock.interview.interview.infra.lock.creation.InterviewCreationUserLock;
 import com.mock.interview.tech.application.helper.TechFinder;
-import com.mock.interview.interview.infra.InterviewRepository;
 import com.mock.interview.tech.infra.TechnicalSubjectsRepository;
 import com.mock.interview.user.domain.exception.UserNotFoundException;
 import com.mock.interview.user.domain.model.Users;
